requirements.txt
requests
kubernetes
click
PyYAML
graphitesend
prometheus_client
setuptools
pytest>=3.6
pytest-cov
pytest-mock
requests-mock
